+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -195,7 +195,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2025-02-19T15:51:29.392Z",
+    "2025-02-24T06:51:40.732Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -229,7 +229,7 @@ export const FileListData$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2025-02-19T15:51:29.392Z"))
+  created: z.date().default(() => new Date("2025-02-24T06:51:40.732Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {